requirements.txt
asgiref==3.7.2 
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django==4.2.2
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.6
redis==5.0.0